Restore original word2med.py file
